Fix a bug
There is a bug in MenuSystem software that causes main menu bar to not update it's transparent parts properly
if background image changes e.g. by moving the background image using arrows, old transparent parts are displayed from the time of first setup.

The fix is a one line of code in the program.