Try to implement opening picture file and displaying it on top of a background image.
This feature was started a long time ago and not finished.
I returned to coding of this project after a long time
so I don't remember everything I wanted to do.

Create gui dialog to choose layer and zoom of image.
The dialog for opening image is not what I would like now.

Add several new classes. Each class has it's own separate file.

It does not work.