Remake displaying dialog window to open picture file using FunnyGui
Because of that remake updating and drawing of GUI widgets
as handling FunnyGUI widgets is different from handling MenuSystem widgets.

It is buggy:

Newline in Label instance is not making newline,
this will not change in FunnyGUI, we have to deal with it another way.

FunnyGUI widgets contained in the Window widget instance
do not react on mouse hovering over it,
it is caused by a bug in FunnyGUI regarding positions of mouse events,
FunnyGUI git commit bcbb9afc4bbc32bae68881471459752921b80bbb.