Add feature create background image filled with a color.
Let the user specify width and height
and a color the picture will be filled with.

Set background image at the start of the application
to be filled with color black and to have width 2000 and height 2000
instead of loading image file from disk.

Enhance the code.
Replace variables starting with 'self.'
with variables without 'self.'
Use positioning with '.rect.topleft='
instead of '.rect.move_ip('.
and adjust their use appropriately.
Add and use methods add_window and remove_window.