Make git include directory "savegames" and enhance FunnyPathGetter usage.
Set caption of FunnyPathGetter windows.
Setting caption is only possible from FunnyPathGetter commit number
7e4f5ff34e6b0b903d3668838c22990570dd9fc0 .

Make git include directory "savegames"
so that it is always present
because I want to count on being there for
savegame files to be saved there and loaded from there.
It is very convenient.
Make PathGetter open that directory
if user wants to load a saved game.
This is important.