Add feature limit selecting to given layers
Add feature limit selectable and highlightable pictures
to those that are within specified layers.

Add feature turn off this behaviour.

Layers can be specified in a GUI dialog.

Layers can be given as digits separated with commas or ranges
or mix of comma separated digits and ranges.
Layers must exist.

Examples of accepted notation
(on condition that the layers with given numbers exist):
1
1,2,3
2,1,3
1-3
1-3,5
0,2-5,7

Examples of not accepted notation:
0-1
1-2
2-3
1-2,2-3
1,,2
3-1