Add file CHANGELOG.md
Some people say it is a good practice and encourage to have and update a changelog file.
The website https://keepachangelog.com/en/1.0.0/ was very helpfull.
I followed advice on the website.

So I plan to update the changelog often.
There is a section Unreleased that should be updated often with all new features changes like feature addition, feature removal, feature depreciation etc.